refactor: 칸반 모달을 칸반 삭제 모달로 이름 변경
칸반 모달을 만들기 위함
