+++ client/src/components/index.ts
@@ -18,4 +18,4 @@ export { default as KanbanColumn } from './KanbanColumn/index';
 export { default as KanbanItem } from './KanbanItem/index';
 export { default as KanbanItemInput } from './KanbanItemInput/index';
 export { default as KanbanAddBtn } from './KanbanAddBtn/index';
-export { default as KanbanModal } from './KanbanModal/index';
+export { default as KanbanDeleteModal } from './KanbanDeleteModal/index';
